Move methods from ValueArrayUtils to SerializerHelper
